Create BasePage, also move ContactMixin to prevent circular dependency
