+++ tbx/core/utils/models.py
@@ -136,6 +136,54 @@ class ColourThemeMixin(models.Model):
             return ColourTheme.NONE
 
 
+class ContactMixin(models.Model):
+    """
+    Provides a `contact` field so that a page can have its own contact
+    in the site-wide footer, instead of the default contact.
+    """
+
+    contact = models.ForeignKey(
+        "people.Contact",
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name="+",
+        help_text="The contact will be applied to this page's footer and all of its "
+        "descendants.\nIf no contact is selected, it will be derived from "
+        "this page's ancestors, eventually falling back to the default contact.",
+    )
+
+    promote_panels = [FieldPanel("contact")]
+
+    class Meta:
+        abstract = True
+
+    @cached_property
+    def footer_contact(self):
+        """
+        Use the page's own contact if set, otherwise, derive the contact from
+        its ancestors, and finally fall back to the default contact.
+
+        NOTE: if, for some reason, a default contact doesn't exist, this will
+        return None, in which case, we'll not display the block in the footer template.
+        """
+        from tbx.people.models import Contact
+
+        if contact := self.contact:
+            return contact
+
+        ancestors = (
+            self.get_ancestors().defer_streamfields().specific().order_by("-depth")
+        )
+        for ancestor in ancestors:
+            if getattr(ancestor, "contact_id", None) is not None:
+                return ancestor.contact
+
+        # _in theory_, there should only be one Contact object with default_contact=True.
+        # (see `tbx.people.models.Contact.save()`)
+        return Contact.objects.filter(default_contact=True).first()
+
+
 class DivisionMixin(models.Model):
     """
     Provides a 'division' field to allow pages to be associated to a Division.
