+++ tbx/services/models.py
@@ -5,49 +5,36 @@ from wagtail.fields import RichTextField
 from wagtail.models import Page
 from wagtail.search import index
 
+from tbx.core.models import BasePage
 from tbx.core.utils.fields import StreamField
 from tbx.core.utils.models import (
     ColourThemeMixin,
+    ContactMixin,
     DivisionMixin,
     NavigationFields,
     SocialFields,
 )
-from tbx.people.models import ContactMixin
 
 from .blocks import ServiceAreaStoryBlock, ServiceStoryBlock
 
 
-class ServicePage(ColourThemeMixin, ContactMixin, SocialFields, NavigationFields, Page):
+class ServicePage(BasePage):
     template = "patterns/pages/service/service_page.html"
 
     intro = RichTextField(blank=True)
     body = StreamField(ServiceStoryBlock())
 
-    content_panels = Page.content_panels + [
+    content_panels = BasePage.content_panels + [
         FieldPanel("intro"),
         FieldPanel("body"),
     ]
 
-    promote_panels = (
-        [
-            MultiFieldPanel(Page.promote_panels, "Common page configuration"),
-        ]
-        + NavigationFields.promote_panels
-        + ColourThemeMixin.promote_panels
-        + ContactMixin.promote_panels
-        + [
-            MultiFieldPanel(SocialFields.promote_panels, "Social fields"),
-        ]
-    )
-
-    search_fields = Page.search_fields + [
+    search_fields = BasePage.search_fields + [
         index.SearchField("body"),
     ]
 
 
-class ServiceAreaPage(
-    ColourThemeMixin, ContactMixin, DivisionMixin, SocialFields, NavigationFields, Page
-):
+class ServiceAreaPage(DivisionMixin, BasePage):
     page_description = "A group of services for a division"
     template = "patterns/pages/service/service_area_page.html"
 
@@ -59,7 +46,7 @@ class ServiceAreaPage(
 
     # Panels
 
-    content_panels = Page.content_panels + [
+    content_panels = BasePage.content_panels + [
         FieldPanel("subtitle"),
         FieldPanel("body"),
     ]
@@ -77,6 +64,6 @@ class ServiceAreaPage(
         ]
     )
 
-    search_fields = Page.search_fields + [
+    search_fields = BasePage.search_fields + [
         index.SearchField("body"),
     ]
